Reorg test helpers
